+++ src/web/middleware/jira/jira-context-symmetric-jwt-auth-middleware.ts
@@ -1,19 +1,7 @@
-import {
-	decodeSymmetric,
-	getAlgorithm,
-	SymmetricAlgorithm,
-} from 'atlassian-jwt';
 import type { NextFunction, Request, RequestHandler, Response } from 'express';
 
-import { verifyExpClaim, verifyQshEqualTo } from './jwt-utils';
-import { CONNECT_JWT_CLAIMS_SCHEMA } from './schemas';
-
-import { isEnumValueOf } from '../../../common/enumUtils';
-import { assertSchema, getLogger } from '../../../infrastructure';
-import { connectInstallationRepository } from '../../../infrastructure/repositories';
+import { jiraInboundAuthService } from '../../../infrastructure/jira/jira-inbound-auth-service';
 import { UnauthorizedError } from '../errors';
-
-const CONTEXT_TOKEN_QSH = 'context-qsh';
 
 /**
  * Authenticates requests using a context symmetric JWT token.
@@ -37,58 +25,12 @@ export const jiraContextSymmetricJwtAuthMiddleware: RequestHandler = (
 		return next(new UnauthorizedError('Missing JWT token.'));
 	}
 
-	void verifyContextSymmetricJwtToken(token)
+	void jiraInboundAuthService
+		.verifyContextSymmetricJwtToken(token)
 		.then(({ connectInstallation, atlassianUserId }) => {
 			res.locals.connectInstallation = connectInstallation;
 			res.locals.atlassianUserId = atlassianUserId;
 			next();
 		})
 		.catch(next);
-};
-
-/**
- * @see https://developer.atlassian.com/cloud/jira/platform/understanding-jwt-for-connect-apps/#decoding-and-verifying-a-jwt-token
- */
-const verifyContextSymmetricJwtToken = async (token: string) => {
-	try {
-		const tokenSigningAlgorithm = getAlgorithm(token) as unknown;
-
-		if (!isEnumValueOf(SymmetricAlgorithm, tokenSigningAlgorithm)) {
-			throw new UnauthorizedError('Unsupported JWT signing algorithm.');
-		}
-
-		// Decode a JWT token without verification.
-		const unverifiedClaims = decodeSymmetric(
-			token,
-			'',
-			tokenSigningAlgorithm,
-			true,
-		) as unknown;
-
-		assertSchema(unverifiedClaims, CONNECT_JWT_CLAIMS_SCHEMA);
-
-		const connectInstallation =
-			await connectInstallationRepository.getByClientKey(unverifiedClaims.iss);
-
-		const verifiedClaims = decodeSymmetric(
-			token,
-			connectInstallation.sharedSecret,
-			tokenSigningAlgorithm,
-		) as unknown;
-
-		assertSchema(verifiedClaims, CONNECT_JWT_CLAIMS_SCHEMA);
-		verifyQshEqualTo(verifiedClaims, CONTEXT_TOKEN_QSH);
-		verifyExpClaim(verifiedClaims);
-
-		return {
-			connectInstallation,
-			atlassianUserId: verifiedClaims.sub!,
-		};
-	} catch (e) {
-		getLogger().warn(e, 'Failed to verify the context symmetric JWT token.');
-
-		if (e instanceof UnauthorizedError) throw e;
-
-		throw new UnauthorizedError('Authentication failed.');
-	}
 };
